Move some files to utils This will break pickle loading To fix, copy go.py to root directory or alternatively delete games.txt
